[FIX] website_event_track_exhibitor: fix the hour displayed in the connect modal
Bug
===

On website, we can join the room of an exhibitor. If we are not in the right
time range a popup is displayed. If the exhibitor room starts at 10h30
it currently displays 10.5 instead.

Technical
=========

The option "float_time" only work in the python templates.
Here, it was used in a JS template.

Task ID-2345894
PR #58516

closes odoo/odoo#58809

X-original-commit: 39513fdec91f9a0ca818e7824db784018709fd1b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>