[FIX] website_event_track_session: jump to unfinished session after an ended talk
This commit adds a condition to the algorithm that redirects the track viewers
to another talk when the one they looked at ended up.
The next talk must be still ongoing. It avoids to redirect to a talk that
started less than 10 minutes ago but that is already done.

Task ID: 2344522

closes odoo/odoo#58836

X-original-commit: e047fc2c7f5d2440adce30ce81a2c10a3b133083
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>