[FIX] web_tour: shield tips design from external css rules
As the tips can be put anywhere in the DOM, external css rules may
target them by mistake. This was for instance discovered on the frontend
where the boxed layout does something like this:
```
    #wrapwrap > * { width: XXX; }
```
which does breaks the tips which would be placed in the #wrapwrap.

This commit solves the problem by forcing rules with !important flags.

closes odoo/odoo#58874

X-original-commit: 5084ea9b4d24c41a8171f5c9ad92f903f1067705
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>