[FIX] web_editor: fix background-image urls becoming absolute on drop
With the recent addition of the parallax option on all snippets,
background-image urls were being converted from relative to absolute on
snippet-drop. This was caused by trying to read the background-image's
url and setting it back onto the target, but there is no way to reliably
read a background-image url in relative form, so it was returned in
absolute form and then written back to the element as is.

This commit fixes that by always converting the absolute URL to relative
when they are on the same origin as the database.

task-2312878

closes odoo/odoo#59113

X-original-commit: 7558dcce791f707e3f3470835133f91ec7ff074f
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>