[FIX] crm: fix checking duplicates on dummy emails
Currently when trying to find duplicates of a lead with an email that
cannot be normalized and without partner, a crash appears as there is
no domain part to form a valid domain.

Fix it by returning a void set if no valid email and no valid partner
is set on the lead as we cannot determine any duplicate based on void
information.

closes odoo/odoo#59147

X-original-commit: eec96bf15680a37933d656d35cdb38f7d2a63dff
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>